requirements.txt
pandas
matplotlib
seaborn
plotly
streamlit
jupyterlab
requests
beautifulsoup4
selenium
python-dotenv
